# frontend/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Governan\u00e7a de Cadastros",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/imports/batches": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Read Batches",
        "operationId": "read_batches_api_v1_imports_batches_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatchListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Post Batch",
        "operationId": "post_batch_api_v1_imports_batches_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_batch_api_v1_imports_batches_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatchProcessingStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/batches/{batch_id}/upload/status": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Read Import Batch Upload Status",
        "operationId": "read_import_batch_upload_status_api_v1_imports_batches__batch_id__upload_status_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatchProcessingStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/batches/{batch_id}": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Read Batch",
        "operationId": "read_batch_api_v1_imports_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatchPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "imports"
        ],
        "summary": "Remove Batch",
        "operationId": "remove_batch_api_v1_imports_batches__batch_id__delete",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatchDeleteStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/batches/{batch_id}/delete/status": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Read Import Batch Delete Status",
        "operationId": "read_import_batch_delete_status_api_v1_imports_batches__batch_id__delete_status_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatchDeleteStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/imports/batches/{batch_id}/row-errors": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Read Batch Row Errors",
        "operationId": "read_batch_row_errors_api_v1_imports_batches__batch_id__row_errors_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRowErrorListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_post_batch_api_v1_imports_batches_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_post_batch_api_v1_imports_batches_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportBatchDeleteStatus": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "processed": {
            "type": "integer",
            "title": "Processed"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "percent": {
            "type": "integer",
            "title": "Percent"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "processed",
          "total",
          "percent",
          "message"
        ],
        "title": "ImportBatchDeleteStatus"
      },
      "ImportBatchDeleteStatusResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ImportBatchDeleteStatus"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ImportBatchDeleteStatusResponse"
      },
      "ImportBatchListData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ImportBatchRead"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "ImportBatchListData"
      },
      "ImportBatchListResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ImportBatchListData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ImportBatchListResponse"
      },
      "ImportBatchPreviewData": {
        "properties": {
          "batch": {
            "$ref": "#/components/schemas/ImportBatchRead"
          },
          "headers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Headers"
          },
          "sample_rows": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Sample Rows"
          }
        },
        "type": "object",
        "required": [
          "batch",
          "headers",
          "sample_rows"
        ],
        "title": "ImportBatchPreviewData"
      },
      "ImportBatchPreviewResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ImportBatchPreviewData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ImportBatchPreviewResponse"
      },
      "ImportBatchProcessingStatus": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "processed": {
            "type": "integer",
            "title": "Processed"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "percent": {
            "type": "integer",
            "title": "Percent"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImportBatchPreviewData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status",
          "processed",
          "total",
          "percent",
          "message"
        ],
        "title": "ImportBatchProcessingStatus"
      },
      "ImportBatchProcessingStatusResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ImportBatchProcessingStatus"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ImportBatchProcessingStatusResponse"
      },
      "ImportBatchRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source_system_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source System Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_hash": {
            "type": "string",
            "title": "File Hash"
          },
          "column_mapping": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column Mapping"
          },
          "source_reference_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "valid_rows": {
            "type": "integer",
            "title": "Valid Rows"
          },
          "invalid_rows": {
            "type": "integer",
            "title": "Invalid Rows"
          },
          "imported_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imported At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_system_id",
          "file_name",
          "file_type",
          "file_hash",
          "column_mapping",
          "source_reference_date",
          "status",
          "total_rows",
          "valid_rows",
          "invalid_rows",
          "imported_at",
          "created_at",
          "updated_at"
        ],
        "title": "ImportBatchRead"
      },
      "ImportRowError": {
        "properties": {
          "row_number": {
            "type": "integer",
            "title": "Row Number"
          },
          "source_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Code"
          },
          "original_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Description"
          },
          "original_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Unit"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "row_number",
          "source_code",
          "original_description",
          "original_unit",
          "issues"
        ],
        "title": "ImportRowError"
      },
      "ImportRowErrorListData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ImportRowError"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "ImportRowErrorListData"
      },
      "ImportRowErrorListResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ImportRowErrorListData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ImportRowErrorListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}